Patch a race condition causing old/unknown handles
There is a race condition where certain remote operations release remote objects close to/before a response containing that remote object gets sent back. This causes a delete command to get recieved before the response does, removing the handle that's needed to understand the object in the following response.

I haven't worked out if it's possible to fix the root cause, but I've added a delay list, so handles get marked for deletion but only removed after a response_timeout period. This means that they should stick around long enough for the response to make it back, except for large/slow edge cases