Serialize/deserialize function partials with a bridged target
Function partials can have a bridged target if you inherit from a remote type and provide local methods. In that case, a call to remote_inheriting_type().local_defined_method() was actually a bridged call to a remote function partial, which in turn bridged calls back to local_defined_method.

To fix this, if a partial has a bridged target, we serialize the target and bound arguments as a PARTIAL type, then create a new, local function partial when it's deserialized. This removes two round-trips for the call.